+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to Hyperspeedcube will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [UNRELEASED]
+
+### Added
+
+- New system for specifying layer masks that change with the number of layers in the puzzle
+
+### Changed
+
+- Default 3D keybinds for <kbd>X</kbd>, <kbd>C</kbd>, <kbd>,</kbd>, and <kbd>.</kbd>
+
 ## [0.7.3] - 2022-08-05
 
 ### Fixed